chore: Version LDAP auth provider
