+++ crates/stackable-operator/src/crd/authentication/ldap/v1alpha1_impl.rs
@@ -1,7 +1,5 @@
 use k8s_openapi::api::core::v1::{Volume, VolumeMount};
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
-use snafu::{ResultExt, Snafu};
+use snafu::{ResultExt as _, Snafu};
 use url::{ParseError, Url};
 
 use crate::{
@@ -9,12 +7,9 @@ use crate::{
         self,
         pod::{container::ContainerBuilder, volume::VolumeMountBuilder, PodBuilder},
     },
-    commons::{
-        networking::HostName,
-        secret_class::{SecretClassVolume, SecretClassVolumeError},
-        tls_verification::{TlsClientDetails, TlsClientDetailsError},
-    },
+    commons::{secret_class::SecretClassVolumeError, tls_verification::TlsClientDetailsError},
     constants::secret::SECRET_BASE_PATH,
+    crd::authentication::ldap::v1alpha1::{AuthenticationProvider, FieldNames},
 };
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
@@ -39,37 +34,6 @@ pub enum Error {
     AddVolumeMounts {
         source: builder::pod::container::Error,
     },
-}
-
-#[derive(
-    Clone, Debug, Deserialize, Eq, Hash, JsonSchema, Ord, PartialEq, PartialOrd, Serialize,
-)]
-#[serde(rename_all = "camelCase")]
-pub struct AuthenticationProvider {
-    /// Host of the LDAP server, for example: `my.ldap.server` or `127.0.0.1`.
-    pub hostname: HostName,
-
-    /// Port of the LDAP server. If TLS is used defaults to 636 otherwise to 389.
-    port: Option<u16>,
-
-    /// LDAP search base, for example: `ou=users,dc=example,dc=org`.
-    #[serde(default)]
-    pub search_base: String,
-
-    /// LDAP query to filter users, for example: `(memberOf=cn=myTeam,ou=teams,dc=example,dc=org)`.
-    #[serde(default)]
-    pub search_filter: String,
-
-    /// The name of the LDAP object fields.
-    #[serde(default)]
-    pub ldap_field_names: FieldNames,
-
-    /// In case you need a special account for searching the LDAP server you can specify it here.
-    bind_credentials: Option<SecretClassVolume>,
-
-    /// Use a TLS connection. If not specified no TLS will be used.
-    #[serde(flatten)]
-    pub tls: TlsClientDetails,
 }
 
 impl AuthenticationProvider {
@@ -168,46 +132,24 @@ impl AuthenticationProvider {
     }
 }
 
-#[derive(
-    Clone, Debug, Deserialize, Eq, Hash, JsonSchema, Ord, PartialEq, PartialOrd, Serialize,
-)]
-#[serde(rename_all = "camelCase")]
-pub struct FieldNames {
-    /// The name of the username field
-    #[serde(default = "FieldNames::default_uid")]
-    pub uid: String,
-    /// The name of the group field
-    #[serde(default = "FieldNames::default_group")]
-    pub group: String,
-    /// The name of the firstname field
-    #[serde(default = "FieldNames::default_given_name")]
-    pub given_name: String,
-    /// The name of the lastname field
-    #[serde(default = "FieldNames::default_surname")]
-    pub surname: String,
-    /// The name of the email field
-    #[serde(default = "FieldNames::default_email")]
-    pub email: String,
-}
-
 impl FieldNames {
-    fn default_uid() -> String {
+    pub(super) fn default_uid() -> String {
         "uid".to_string()
     }
 
-    fn default_group() -> String {
+    pub(super) fn default_group() -> String {
         "memberof".to_string()
     }
 
-    fn default_given_name() -> String {
+    pub(super) fn default_given_name() -> String {
         "givenName".to_string()
     }
 
-    fn default_surname() -> String {
+    pub(super) fn default_surname() -> String {
         "sn".to_string()
     }
 
-    fn default_email() -> String {
+    pub(super) fn default_email() -> String {
         "mail".to_string()
     }
 }
@@ -227,6 +169,7 @@ impl Default for FieldNames {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::commons::secret_class::SecretClassVolume;
 
     #[test]
     fn minimal() {
